Move fp16 tests to existing file and add block comments
